Conflicts:
	WebApplication/DTO/src/main/java/com/thesischecker/dto/UserEntity.java
	WebApplication/DTO/src/main/resources/hibernate.cfg.xml
	WebApplication/app-web/src/main/resources/database.properties
	WebApplication/app-web/src/main/resources/spring/applicationContext.xml
	WebApplication/app-web/src/main/webapp/WEB-INF/web.xml
	WebApplication/pom.xml

+++ WebApplication/app-web/src/main/java/com/thesischecker/app/AppServlet.java
@@ -31,7 +31,7 @@ public class AppServlet implements HttpRequestHandler {
     public void handleRequest(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
         UserEntity userEntity = new UserEntity();
-        userEntity.setId(2);
+        userEntity.setId((long) 2);
         userEntity.setEmail("michal.szczygiel@wp.pl");
 
         try {
